// apps/admin/src/app/store/user/user.types.ts
export enum USER_ACTION_TYPES {
  SIGN_IN_START = "user/SIGN_IN_START",
  SET_CURRENT_USER = "user/SET_CURRENT_USER",
  CHECK_USER_SESSION = "user/CHECK_USER_SESSION",
  CHECK_USER_SESSION_FAILED = "user/CHECK_USER_SESSION_FAILED",
  CHECK_USER_SESSION_SUCCESS = "user/CHECK_USER_SESSION_SUCCESS",
  SIGN_IN_SUCCESS = "user/SIGN_IN_SUCCESS",
  SIGN_IN_FAILED = "user/SIGN_IN_FAILED",
  SIGN_OUT_START = "user/SIGN_OUT_START",
  SIGN_OUT_SUCCESS = "user/SIGN_OUT_SUCCESS",
  SIGN_OUT_FAILED = "user/SIGN_OUT_FAILED",
}